+++ src/Sportle/Sportle.Web/Controllers/LeaguesController.cs
@@ -136,7 +136,7 @@ namespace Sportle.Web.Controllers
             return View(model);
         }
 
-        private static int GetUserScore(SportleDbContext context, IdentityUser user, List<Guid> eventIds)
+        private static double GetUserScore(SportleDbContext context, IdentityUser user, List<Guid> eventIds)
         {
             if (eventIds.Count == 0)
                 return 0;
